+++ tests/dummy/app/app.js
@@ -5,11 +5,7 @@ import config from './config/environment';
 
 import 'npm:../../lib/htmlbars-plugin';
 
-let App;
-
-Ember.MODEL_FACTORY_INJECTIONS = true;
-
-App = Ember.Application.extend({
+const App = Ember.Application.extend({
   modulePrefix: config.modulePrefix,
   podModulePrefix: config.podModulePrefix,
   Resolver